Move password prompt to platform utils
