+++ package.json
@@ -99,7 +99,6 @@
     "open": "7.1.0",
     "papaparse": "5.2.0",
     "rxjs": "6.6.2",
-    "sweetalert2": "10.15.3",
     "tldjs": "2.3.1",
     "zone.js": "0.9.1",
     "zxcvbn": "4.4.2"
