+++ src/abstractions/platformUtils.service.ts
@@ -26,6 +26,7 @@ export abstract class PlatformUtilsService {
         options?: any) => void;
     showDialog: (body: string, title?: string, confirmText?: string, cancelText?: string,
         type?: string, bodyIsHtml?: boolean) => Promise<boolean>;
+    showPasswordDialog: (body: string, title: string, passwordValidation: (value: string) => Promise<boolean>);
     isDev: () => boolean;
     isSelfHost: () => boolean;
     copyToClipboard: (text: string, options?: any) => void;
